c_check: TODO: fix `get_library_include`
ah, another day, another package that the debian packaging team mangle
beyond repair.  this time: libclang.so is being renamed within the
significant part of the library name (libclang.so => libclang-7.so).
the top-quality trash i expect from a debian package!

a better way to do this is to totally separate the llvm sysroots, and
just point into those -- and the debian folks have sort of done that,
but the links are backwards: the shared object is in /usr/lib/$ARCH and
/usr/lib/llvm-7/lib/libclang.so links to /usr/lib/$ARCH

but an even better way is just to ask `clang' where its libclang is:

    :cs1521[19T3]@zappa: clang -print-search-dirs
    programs: =/usr/bin:/usr/lib/llvm-7/bin:/usr/bin/../lib/gcc/x86_64-linux-gnu/8/../../../../x86_64-linux-gnu/bin
    libraries: =/usr/lib/llvm-7/lib/clang/7.0.1:/usr/bin/../lib/gcc/x86_64-linux-gnu/8:/usr/bin/../lib/gcc/x86_64-linux-gnu/8/../../../x86_64-linux-gnu:/lib/x86_64-linux-gnu:/lib/../lib64:/usr/lib/x86_64-linux-gnu:/usr/bin/../lib/gcc/x86_64-linux-gnu/8/../../..:/usr/lib/llvm-7/bin/../lib:/lib:/usr/lib
    :cs1521[19T3]@zappa: ls -F /usr/lib/llvm-7/bin/../lib/libclang*
    /usr/lib/llvm-7/bin/../lib/libclang-7.so.1@
    /usr/lib/llvm-7/bin/../lib/libclang.so.1@

i think iterating through each of the library directories that clang
returns would probably do to find it, and with the name `libclang.so.1'
as would be expected based on every other llvm insatll i've ever seen